Initial Commit
This push is basically the same as that of the "SME Playground" workspace on my desk at the time of creation of this repo.
The Eclipse and Visual Studio 2010 projects build and are debuggable (meaning, single-steppable).
No attention has been paid whatsoever to Doxygen projects nor any other documentation, so treat it all as suspect.
No attempt has been made to open the component-level projects for lib, arch, etc.