Refactor EvQX struct, methods. Added Cwsw_EvQX__SetEvHandler().
The old format of an EvQX object was to contain references to a normal EvQ object and a reference to an Event Handler Association object.
This introduces risk because of the similarity in the method calls between EvQ and EvQX objects: passing an EvQ object to an EvQX method, or vice versa, would be catastrophic.

This push makes the EvQX object to be a superset of an EvQ object, to reduce risk when mixing object type and method calls.

Updated copyright block.